requirements.txt
Django==1.4.1
pil
south
django-cms
django-tinymce
django-reversion
django-gravatar
django-registration
django-google-analytics
django-taggit
